major update
Lacks a lot of commits but this provide the rewrite a lot of functions
move almost all function to the frontend (thank gpt for getting me out of that dead end of trying to make cli tui and web-ui working on the same base). This will come later.
The program start to work as intended.